@@ -3,7 +3,7 @@ import 'package:xaycc_portfolio/pages/home_page.dart';
 import 'package:xaycc_portfolio/pages/particle_background.dart';
 
 class PortfolioApp extends StatelessWidget {
-  const PortfolioApp({Key key}) : super(key: key);
+  const PortfolioApp({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
