@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:xaycc_portfolio/ui/anim/particle/particle.dart';
 
 class DisplayPoints extends CustomPainter {
-  final List<Particle> particlesList;
+  final List<Particle>? particlesList;
 
   DisplayPoints({
     this.particlesList,
@@ -10,12 +10,16 @@ class DisplayPoints extends CustomPainter {
 
   @override
   void paint(Canvas canvas, Size size) {
+    if (particlesList == null) return;
+
+    var particles = particlesList as List<Particle>;
+
     Paint line = Paint();
     line.strokeCap = StrokeCap.round;
-    line.color = particlesList.elementAt(0).color;
+    line.color = particles.elementAt(0).color;
 
     // Draw all the particles
-    for (var particle in particlesList) {
+    for (var particle in particles) {
       line.strokeWidth = particle.size;
       Offset center = Offset(particle.xCoor, particle.yCoor);
       line.style = PaintingStyle.fill;
@@ -23,10 +27,10 @@ class DisplayPoints extends CustomPainter {
     }
 
     // Draw the connect line
-    for (int i = 0; i < this.particlesList.length; i++) {
-      for (int j = i + 1; j < this.particlesList.length; j++) {
-        Particle particle = this.particlesList.elementAt(i);
-        Particle anotherParticle = this.particlesList.elementAt(j);
+    for (int i = 0; i < particles.length; i++) {
+      for (int j = i + 1; j < particles.length; j++) {
+        Particle particle = particles.elementAt(i);
+        Particle anotherParticle = particles.elementAt(j);
         if (particle.isNear(anotherParticle)) {
           Offset firstParticle = Offset(particle.xCoor, particle.yCoor);
           Offset secondParticle =
