@@ -22,13 +22,12 @@ import 'package:flutter/scheduler.dart';
 /// [startTimeSimulationTicks] property.)
 class Rendering extends StatefulWidget {
   final Widget Function(BuildContext context, Duration timeElapsed) builder;
-  final Function(Duration timeElapsed) onTick;
+  final Function(Duration timeElapsed)? onTick;
   final Duration startTime;
   final int startTimeSimulationTicks;
 
-  Rendering(
-      {this.builder,
-      this.onTick,
+  Rendering(this.builder,
+      {this.onTick,
       this.startTime = Duration.zero,
       this.startTimeSimulationTicks = 20})
       : assert(builder != null, "Builder needs to defined.");
@@ -39,7 +38,7 @@ class Rendering extends StatefulWidget {
 
 class _RenderingState extends State<Rendering>
     with SingleTickerProviderStateMixin {
-  Ticker _ticker;
+  late Ticker _ticker;
   Duration _timeElapsed = Duration(milliseconds: 0);
 
   @override
@@ -57,7 +56,7 @@ class _RenderingState extends State<Rendering>
 
   void _onRender(Duration effectiveElapsed) {
     if (widget.onTick != null) {
-      widget.onTick(effectiveElapsed);
+      widget.onTick!(effectiveElapsed);
     }
     setState(() {
       _timeElapsed = effectiveElapsed;
@@ -72,7 +71,7 @@ class _RenderingState extends State<Rendering>
                     i /
                     widget.startTimeSimulationTicks)
                 .round());
-        widget.onTick(simulatedTime);
+        widget.onTick!(simulatedTime);
       });
     }
   }
@@ -158,7 +157,7 @@ class MultiTrackTween extends Animatable<Map<String, dynamic>> {
       final sequence = TweenSequence(sequenceItems);
 
       _tracksToTween[track.name] =
-          _TweenData(tween: sequence, maxTime: trackDuration / _maxDuration);
+          _TweenData(sequence, trackDuration / _maxDuration);
     });
   }
 
@@ -209,7 +208,7 @@ class Track<T> {
   ///
   /// Optionally you can set a named parameter [curve] that applies an easing
   /// curve to the tween.
-  Track<T> add(Duration duration, Animatable<T> tween, {Curve curve}) {
+  Track<T> add(Duration duration, Animatable<T> tween, {Curve? curve}) {
     items.add(_TrackItem(duration, tween, curve: curve));
     return this;
   }
@@ -217,9 +216,9 @@ class Track<T> {
 
 class _TrackItem<T> {
   final Duration duration;
-  Animatable<T> tween;
+  late Animatable<T> tween;
 
-  _TrackItem(this.duration, Animatable<T> _tween, {Curve curve})
+  _TrackItem(this.duration, Animatable<T> _tween, {Curve? curve})
       : assert(duration != null, "Please set a duration."),
         assert(_tween != null, "Please set a tween.") {
     if (curve != null) {
@@ -231,10 +230,13 @@ class _TrackItem<T> {
 }
 
 class _TweenData<T> {
-  final Animatable<T> tween;
-  final double maxTime;
+  late final Animatable<T> tween;
+  late final double maxTime;
 
-  _TweenData({this.tween, this.maxTime});
+  _TweenData(Animatable<T> tween, double maxTime) {
+    this.maxTime = maxTime;
+    this.tween = tween;
+  }
 }
 
 /// Playback tell the controller of the animation what to do.
@@ -302,29 +304,30 @@ enum Playback {
 ///
 class ControlledAnimation<T> extends StatefulWidget {
   final Playback playback;
-  final Animatable<T> tween;
+  final Animatable<T>? tween;
   final Curve curve;
-  final Duration duration;
-  final Duration delay;
-  final Widget Function(BuildContext buildContext, T animatedValue) builder;
-  final Widget Function(BuildContext, Widget child, T animatedValue)
+  final Duration? duration;
+  final Duration? delay;
+  late final Widget Function(BuildContext buildContext, T animatedValue)
+      builder;
+  final Widget Function(BuildContext, Widget child, T animatedValue)?
       builderWithChild;
-  final Widget child;
-  final AnimationStatusListener animationControllerStatusListener;
+  final Widget? child;
+  final AnimationStatusListener? animationControllerStatusListener;
   final double startPosition;
 
   ControlledAnimation(
-      {this.playback = Playback.PLAY_FORWARD,
+      Widget Function(BuildContext buildContext, T animatedValue) builder,
+      {this.child,
       this.tween,
-      this.curve = Curves.linear,
+      this.builderWithChild,
       this.duration,
       this.delay,
-      this.builder,
-      this.builderWithChild,
-      this.child,
+      this.playback = Playback.PLAY_FORWARD,
+      this.curve = Curves.linear,
       this.animationControllerStatusListener,
       this.startPosition = 0.0,
-      Key key})
+      Key? key})
       : assert(duration != null,
             "Please set property duration. Example: Duration(milliseconds: 500)"),
         assert(tween != null,
@@ -338,7 +341,9 @@ class ControlledAnimation<T> extends StatefulWidget {
             startPosition >= 0 && startPosition <= 1,
             "The property startPosition "
             "must have a value between 0.0 and 1.0."),
-        super(key: key);
+        super(key: key) {
+    this.builder = builder;
+  }
 
   @override
   _ControlledAnimationState<T> createState() => _ControlledAnimationState<T>();
@@ -346,8 +351,8 @@ class ControlledAnimation<T> extends StatefulWidget {
 
 class _ControlledAnimationState<T> extends State<ControlledAnimation>
     with SingleTickerProviderStateMixin {
-  AnimationController _controller;
-  Animation<T> _animation;
+  late AnimationController _controller;
+  late Animation<dynamic> _animation;
   bool _isDisposed = false;
   bool _waitForDelay = true;
   bool _isCurrentlyMirroring = false;
@@ -360,12 +365,13 @@ class _ControlledAnimationState<T> extends State<ControlledAnimation>
       })
       ..value = widget.startPosition;
 
-    _animation = widget.tween
+    _animation = widget.tween!
         .chain(CurveTween(curve: widget.curve))
         .animate(_controller);
 
     if (widget.animationControllerStatusListener != null) {
-      _controller.addStatusListener(widget.animationControllerStatusListener);
+      _controller.addStatusListener(
+          widget.animationControllerStatusListener as AnimationStatusListener);
     }
 
     initialize();
@@ -374,7 +380,7 @@ class _ControlledAnimationState<T> extends State<ControlledAnimation>
 
   void initialize() async {
     if (widget.delay != null) {
-      await Future.delayed(widget.delay);
+      await Future.delayed(widget.delay as Duration);
     }
     _waitForDelay = false;
     executeInstruction();
@@ -425,7 +431,8 @@ class _ControlledAnimationState<T> extends State<ControlledAnimation>
     if (widget.builder != null) {
       return widget.builder(context, _animation.value);
     } else if (widget.builderWithChild != null && widget.child != null) {
-      return widget.builderWithChild(context, widget.child, _animation.value);
+      return widget.builderWithChild!(
+          context, widget.child as Widget, _animation.value);
     }
     _controller.stop(canceled: true);
     throw FlutterError(
@@ -442,18 +449,22 @@ class _ControlledAnimationState<T> extends State<ControlledAnimation>
 }
 
 class AnimationProgress {
-  final Duration duration;
-  final Duration startTime;
+  final Duration? duration;
+  final Duration? startTime;
 
   /// Creates an [AnimationProgress].
   AnimationProgress({this.duration, this.startTime})
       : assert(duration != null, "Please specify an animation duration."),
-        assert(
-            startTime != null, "Please specify a start time of the animation.");
+        assert(startTime != null,
+            "Please specify a start time of the animation.") {}
 
   /// Queries the current progress value based on the specified [startTime] and
   /// [duration] as a value between `0.0` and `1.0`. It will automatically
   /// clamp values this interval to fit in.
-  double progress(Duration time) => max(0.0,
-      min((time - startTime).inMilliseconds / duration.inMilliseconds, 1.0));
+  double progress(Duration time) => max(
+      0.0,
+      min(
+          (time - (startTime as Duration)).inMilliseconds /
+              (duration as Duration).inMilliseconds,
+          1.0));
 }
