@@ -1,13 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:url_launcher/url_launcher.dart';
 import 'package:xaycc_portfolio/data/const.dart';
 import 'package:xaycc_portfolio/ui/widgets/NavContactButtonWidget.dart';
 
 class MainMenuWidget extends StatelessWidget {
-  final List<Widget> navButtons;
-  final List<Widget> navTopButtons;
+  final List<Widget>? navButtons;
+  final List<Widget>? navTopButtons;
 
-  const MainMenuWidget({Key key, this.navButtons, this.navTopButtons})
+  const MainMenuWidget({Key? key, this.navButtons, this.navTopButtons})
       : super(key: key);
 
   Widget build(BuildContext context) {
@@ -20,16 +19,17 @@ class MainMenuWidget extends StatelessWidget {
             text: Const.phone,
             textColor: Colors.white,
             padding: EdgeInsets.fromLTRB(32, 24, 32, 0),
-            onPressed: () => launch("tel:${Const.phone}")),
+            onPressed: {}),
         NavContactButtonWidget(
-            icon: Icons.email_outlined,
-            text: Const.email,
-            textColor: Colors.white,
-            padding: EdgeInsets.fromLTRB(32, 0, 32, 0),
-            onPressed: () => launch("mailto:${Const.email}")),
+          icon: Icons.email_outlined,
+          text: Const.email,
+          textColor: Colors.white,
+          padding: EdgeInsets.fromLTRB(32, 0, 32, 0),
+          onPressed: {},
+        ),
         ListView(
           padding: const EdgeInsets.all(20),
-          children: navButtons,
+          children: navButtons ?? [],
         )
       ],
     );
