@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:xaycc_portfolio/infrastructure/tools/responsive_widget.dart';
 
 class MainTopHeaderWidget extends StatelessWidget {
-  final List<Widget> navButtons;
+  final List<Widget>? navButtons;
 
-  const MainTopHeaderWidget({Key key, this.navButtons}) : super(key: key);
+  const MainTopHeaderWidget({Key? key, this.navButtons}) : super(key: key);
 
   Widget build(BuildContext context) {
     return ResponsiveWidget(
@@ -15,7 +15,7 @@ class MainTopHeaderWidget extends StatelessWidget {
         crossAxisAlignment: CrossAxisAlignment.end,
         children: <Widget>[
           Row(
-            children: navButtons,
+            children: navButtons ?? [],
           )
         ],
       ),
